+++ examples/hcs_wizard.py
@@ -10,7 +10,7 @@ except ImportError:
 from useq import GridRowsColumns, WellPlatePlan
 
 from pymmcore_widgets.hcs import HCSWizard
-from pymmcore_widgets.hcs._plate_model import load_database
+from pymmcore_widgets.hcs._util import load_database
 
 database_path = Path(__file__).parent.parent / "tests" / "plate_database_for_tests.json"
 database = load_database(database_path)
