+++ src/pymmcore_widgets/hcs/_util.py
@@ -1,7 +1,10 @@
 from __future__ import annotations
 
+import json
 import math
+import warnings
 from itertools import product
+from pathlib import Path
 from typing import TYPE_CHECKING
 
 import numpy as np
@@ -11,15 +14,17 @@ from qtpy.QtWidgets import (
     QGraphicsView,
     QWidget,
 )
+from useq import WellPlate
 
 from ._graphics_items import _WellGraphicsItem
 
 if TYPE_CHECKING:
     from qtpy.QtGui import QBrush, QPen, QResizeEvent
-    from useq import WellPlate
     from useq._grid import GridPosition
 
     from ._graphics_items import Well
+
+DEFAULT_PLATE_DB_PATH = Path(__file__).parent / "default_well_plate_database.json"
 
 
 class _ResizingGraphicsView(QGraphicsView):
@@ -44,7 +49,7 @@ def draw_plate(
     """Draw all wells of the plate."""
     # setting a custom well size in scene px. Using 10 times the well size in mm
     # gives a good resolution in the viewer.
-    well_size_x, well_size_y = plate.well_size  # type: ignore  # need my useq PR
+    well_size_x, well_size_y = plate.well_size
     well_spacing_x, well_spacing_y = plate.well_spacing
 
     scene.clear()
@@ -185,15 +190,42 @@ def _top_left(points: list[GridPosition], top_x: float, top_y: float) -> GridPos
     )[0]
 
 
-# def _furthest_point_from_center(points: list[GridPosition]) -> GridPosition | None:
-#     """Find the furthest point from the center (0, 0)."""
-#     max_distance: float = 0.0
-#     furthest_point: GridPosition | None = None
+def load_database(path: str | Path = DEFAULT_PLATE_DB_PATH) -> dict[str, WellPlate]:
+    """Load the database of well plates contained in database_path.
 
-#     for point in points:
-#         distance = math.sqrt(point.x**2 + point.y**2)
-#         if distance > max_distance:
-#             max_distance = distance
-#             furthest_point = point
+    If an error occurs, it will load the default database.
 
-#     return furthest_point
+    Database must be a JSON file with this structure:
+
+    "6-well": {
+        "rows": 2,
+        "columns": 3,
+        "well_spacing": 39.12,
+        "well_size": 34.8
+        "circular_wells": true"  # optional, by default True
+    },
+    ...
+    """
+    if Path(path) == DEFAULT_PLATE_DB_PATH:
+        with open(path) as f:
+            return {k: WellPlate(**v, name=k) for k, v in json.load(f).items()}
+    try:
+        with open(path) as f:
+            return {k: WellPlate(**v, name=k) for k, v in json.load(f).items()}
+    except Exception as e:
+        warnings.warn(
+            f"Error loading well plate database: {e}."
+            f"Loading default database from {DEFAULT_PLATE_DB_PATH}.",
+            stacklevel=2,
+        )
+        with open(DEFAULT_PLATE_DB_PATH) as f:
+            return {k: WellPlate(**v, name=k) for k, v in json.load(f).items()}
+
+
+def save_database(database: dict[str, WellPlate], database_path: Path | str) -> None:
+    """Save the database of well plates to database_path.
+
+    The database will be saved as a JSON file.
+    """
+    with open(Path(database_path), "w") as f:
+        json.dump({k: v.model_dump_json() for k, v in database.items()}, f, indent=4)
